fixed bug: final sample of prev note was first sample of new note
This repetition of a single sample caused audible artifacts, so
I fixed it.  However, this had the unfortunate side effect of
breaking the "ufo" sound in the test script.  Especially when played
in a triangle wave style, it got most of its interesting traits from
the bug.  Now instead of a UFO sound, it's just a boring slow vibrato.